Handle exceptions and decode pygments highlight result
This patch clears Python exceptions (so they aren't thrown in the
following invocations to the interpreter) and properly decodes the
pygments.highlight() result. The previous implementation of using
PyString_AsString() used the default codec (which on my system was
'ascii'); this caused problems for strings with outside of the ASCII
code pages. The implementation now properly decodes the unicode object
to a bytes object using utf8.